+++ tests/conftest.py
@@ -1,4 +1,8 @@
+import os
+
 import pytest
+
+from pushsource import Source
 
 
 @pytest.fixture
@@ -6,7 +10,7 @@ def fixture_unsorted_files_json():
     json = [
         {
             "type": "product",
-            "state": "create",
+            "action": "create",
             "metadata": {
                 "name": "Cloud: Ansible NEW Test",
                 "productCode": "AnsibleNewTest",
@@ -20,7 +24,7 @@ def fixture_unsorted_files_json():
         },
         {
             "type": "product_version",
-            "state": "create",
+            "action": "create",
             "metadata": {
                 "productName": "Cloud: Ansible NEW Test",
                 "productCode": "AnsibleNewTest",
@@ -36,7 +40,7 @@ def fixture_unsorted_files_json():
         },
         {
             "type": "product",
-            "state": "delete",
+            "action": "delete",
             "metadata": {
                 "name": "Cloud: Ansible NEW Test 2",
                 "productCode": "AnsibleNewTest",
@@ -50,7 +54,7 @@ def fixture_unsorted_files_json():
         },
         {
             "type": "product_version",
-            "state": "delete",
+            "action": "delete",
             "metadata": {
                 "productName": "Cloud: Ansible NEW Test 2",
                 "productCode": "AnsibleNewTest",
@@ -66,7 +70,7 @@ def fixture_unsorted_files_json():
         },
         {
             "type": "file",
-            "state": "create",
+            "action": "create",
             "metadata": {
                 "productName": "Cloud: Ansible NEW TEST",
                 "productCode": "AnsibleNewTest",
@@ -86,7 +90,7 @@ def fixture_unsorted_files_json():
         },
         {
             "type": "file",
-            "state": "delete",
+            "action": "delete",
             "metadata": {
                 "productName": "Cloud: Ansible NEW TESTTEST",
                 "productCode": "AnsibleNewTest",
@@ -113,7 +117,7 @@ def fixture_sorted_files_json():
     json = [
         {
             "type": "product",
-            "state": "create",
+            "action": "create",
             "metadata": {
                 "name": "Cloud: Ansible NEW Test",
                 "productCode": "AnsibleNewTest",
@@ -127,7 +131,7 @@ def fixture_sorted_files_json():
         },
         {
             "type": "product_version",
-            "state": "create",
+            "action": "create",
             "metadata": {
                 "productName": "Cloud: Ansible NEW Test",
                 "productCode": "AnsibleNewTest",
@@ -143,7 +147,7 @@ def fixture_sorted_files_json():
         },
         {
             "type": "file",
-            "state": "create",
+            "action": "create",
             "metadata": {
                 "productName": "Cloud: Ansible NEW TEST",
                 "productCode": "AnsibleNewTest",
@@ -163,7 +167,7 @@ def fixture_sorted_files_json():
         },
         {
             "type": "file",
-            "state": "delete",
+            "action": "delete",
             "metadata": {
                 "productName": "Cloud: Ansible NEW TESTTEST",
                 "productCode": "AnsibleNewTest",
@@ -183,7 +187,7 @@ def fixture_sorted_files_json():
         },
         {
             "type": "product_version",
-            "state": "delete",
+            "action": "delete",
             "metadata": {
                 "productName": "Cloud: Ansible NEW Test 2",
                 "productCode": "AnsibleNewTest",
@@ -199,7 +203,7 @@ def fixture_sorted_files_json():
         },
         {
             "type": "product",
-            "state": "delete",
+            "action": "delete",
             "metadata": {
                 "name": "Cloud: Ansible NEW Test 2",
                 "productCode": "AnsibleNewTest",
@@ -219,7 +223,7 @@ def fixture_sorted_files_json():
 def create_product_data():
     return {
         "type": "product",
-        "state": "create",
+        "action": "create",
         "metadata": {
             "name": "Cloud: Ansible NEW Test",
             "productCode": "AnsibleNewTest",
@@ -237,7 +241,7 @@ def create_product_data():
 def create_product2_data():
     return {
         "type": "product",
-        "state": "create",
+        "action": "create",
         "metadata": {
             "name": "Cloud: Ansible NEW Test 1",
             "productCode": "AnsibleNewTest 1",
@@ -255,7 +259,7 @@ def create_product2_data():
 def update_product_data():
     return {
         "type": "product",
-        "state": "update",
+        "action": "update",
         "metadata": {
             "name": "Cloud: Ansible NEW Test",
             "productCode": "AnsibleNewTest",
@@ -273,7 +277,7 @@ def update_product_data():
 def delete_product_data():
     return {
         "type": "product",
-        "state": "delete",
+        "action": "delete",
         "metadata": {
             "name": "Cloud: Ansible NEW Test 1",
             "productCode": "AnsibleNewTest 1",
@@ -291,7 +295,7 @@ def delete_product_data():
 def create_version_data():
     return {
         "type": "product_version",
-        "state": "create",
+        "action": "create",
         "metadata": {
             "productName": "Cloud: Ansible NEW Test",
             "productCode": "AnsibleNewTest",
@@ -311,7 +315,7 @@ def create_version_data():
 def update_version_data():
     return {
         "type": "product_version",
-        "state": "update",
+        "action": "update",
         "metadata": {
             "productName": "Cloud: Ansible NEW Test",
             "productCode": "AnsibleNewTest",
@@ -331,7 +335,7 @@ def update_version_data():
 def create_version2_data():
     return {
         "type": "product_version",
-        "state": "create",
+        "action": "create",
         "metadata": {
             "productName": "Cloud: Ansible NEW Test",
             "productCode": "AnsibleNewTest",
@@ -351,7 +355,7 @@ def create_version2_data():
 def create_version_without_product():
     return {
         "type": "product_version",
-        "state": "create",
+        "action": "create",
         "metadata": {
             "productName": "InvalidProduct",
             "productCode": "InvalidProduct",
@@ -371,7 +375,7 @@ def create_version_without_product():
 def delete_version():
     return {
         "type": "product_version",
-        "state": "delete",
+        "action": "delete",
         "metadata": {
             "productName": "Cloud: Ansible NEW Test",
             "productCode": "AnsibleNewTest",
@@ -391,7 +395,7 @@ def delete_version():
 def create_file_data():
     return {
         "type": "file",
-        "state": "create",
+        "action": "create",
         "metadata": {
             "productName": "Cloud: Ansible NEW Test",
             "productCode": "AnsibleNewTest",
@@ -415,7 +419,7 @@ def create_file_data():
 def create_file_without_product():
     return {
         "type": "file",
-        "state": "create",
+        "action": "create",
         "metadata": {
             "productName": "Invalid product name",
             "productCode": "InvalidCode",
@@ -439,7 +443,7 @@ def create_file_without_product():
 def create_file_without_version():
     return {
         "type": "file",
-        "state": "create",
+        "action": "create",
         "metadata": {
             "productName": "Cloud: Ansible NEW Test",
             "productCode": "AnsibleNewTest",
@@ -463,7 +467,7 @@ def create_file_without_version():
 def create_file2_data():
     return {
         "type": "file",
-        "state": "create",
+        "action": "create",
         "metadata": {
             "productName": "Cloud: Ansible NEW Test",
             "productCode": "AnsibleNewTest",
@@ -487,7 +491,7 @@ def create_file2_data():
 def update_file_data():
     return {
         "type": "file",
-        "state": "update",
+        "action": "update",
         "metadata": {
             "productName": "Cloud: Ansible NEW Test",
             "productCode": "AnsibleNewTest",
@@ -511,7 +515,7 @@ def update_file_data():
 def delete_file_data():
     return {
         "type": "file",
-        "state": "delete",
+        "action": "delete",
         "metadata": {
             "productName": "Cloud: Ansible NEW Test",
             "productCode": "AnsibleNewTest",
@@ -529,3 +533,133 @@ def delete_file_data():
             "size": None,
         },
     }
+
+
+@pytest.fixture
+def yml_json_data():
+    # This data will be used to verify the hybrid model of yml structure.
+    # The yml file can accept both linear and nested yml data structure.
+    # The nested yml data will get converted to linear format inorder to process
+
+    json = [
+        {
+            "type": "product",
+            "action": "create",
+            "metadata": {
+                "name": "Product_Name_1",
+                "productCode": "Product_code_1",
+                "homepage": "https://developers.redhat.com/products/codeready-containers/overview/",
+                "downloadpage": "https://developers.redhat.com/products/codeready-containers/download/",
+                "thankYouPage": "https://test.com/",
+                "thankYouTimeout": 5,
+                "eloquaCode": "FAKECODEID1234",
+            },
+        },
+        {
+            "type": "product_version",
+            "action": "create",
+            "metadata": {
+                "versionName": "3.4.0",
+                "masterProductVersion": None,
+                "releaseDate": "2022-25-05",
+                "ga": True,
+                "hidden": False,
+                "invisible": False,
+                "trackingDisabled": False,
+                "termsAndConditions": "Anonymous Download",
+                "productName": "Product_Name_1",
+                "productCode": "Product_code_1",
+            },
+        },
+        {
+            "type": "file",
+            "action": "create",
+            "metadata": {
+                "downloadURL": "/content/origin/files/AnsibleNewTest_1/",
+                "label": "Checksum File",
+                "shortURL": "/pub/openshift-v1/clients/sha256sum.txt",
+                "description": "Red Hat OpenShift Local",
+                "differentProductThankYouPage": None,
+                "order": 10,
+                "productName": "Product_Name_1",
+                "productCode": "Product_code_1",
+                "productVersionName": "3.4.0",
+            },
+        },
+        {
+            "type": "file",
+            "action": "create",
+            "metadata": {
+                "downloadURL": "/content/origin/files/AnsibleNewTest_2/",
+                "label": "Release Info",
+                "shortURL": "/pub/openshift-v2/clients/sha256sum.txt",
+                "description": "Red Hat OpenShift Local",
+                "differentProductThankYouPage": None,
+                "order": 20,
+                "productName": "Product_Name_1",
+                "productCode": "Product_code_1",
+                "productVersionName": "3.4.0",
+            },
+        },
+        {
+            "type": "product",
+            "action": "create",
+            "metadata": {
+                "name": "Cloud: Ansible NEW Test 1",
+                "productCode": "AnsibleNewTest",
+                "homepage": "https://test.com/",
+                "downloadpage": "https://test.com/",
+                "thankYouPage": "https://test.com/",
+                "eloquaCode": "NOT_SET",
+                "featuredArtifactType": "Server",
+                "thankYouTimeout": 5,
+            },
+        },
+        {
+            "type": "product_version",
+            "action": "create",
+            "metadata": {
+                "productName": "Cloud: Ansible NEW Test 1",
+                "productCode": "AnsibleNewTest",
+                "versionName": "AnsibleNewTestVersion 1",
+                "ga": True,
+                "masterProductVersion": None,
+                "termsAndConditions": "Anonymous Download",
+                "trackingDisabled": False,
+                "hidden": False,
+                "invisible": False,
+                "releaseDate": "2022-25-05",
+            },
+        },
+        {
+            "type": "file",
+            "action": "create",
+            "metadata": {
+                "productName": "Cloud: Ansible NEW Test 1",
+                "productCode": "AnsibleNewTest",
+                "productVersionName": "AnsibleNewTestVersion 1",
+                "description": "Red Hat OpenShift Local Sandbox Test",
+                "label": "Checksum File Sandbox Test",
+                "order": 0,
+                "hidden": False,
+                "invisible": False,
+                "type": "FILE",
+                "differentProductThankYouPage": None,
+                "downloadURL": "/content/origin/test",
+                "shortURL": "/pub-1/openshift-v4/test",
+                "md5": None,
+                "size": None,
+            },
+        },
+    ]
+    return json
+
+
+def test_staging_dir():
+    return os.path.join(os.path.dirname(__file__), "test_data/test_staging_dir")
+
+
+@pytest.fixture()
+def fixture_source_stage(request):
+    yield Source.register_backend("stage", lambda: request.param)
+    Source.reset()