+++ tests/test_utils.py
@@ -1,5 +1,5 @@
 import os
-from pubtools._content_gateway.utils import yaml_parser, validate_data, sort_items
+from pubtools._content_gateway.utils import yaml_parser, validate_data, sort_items, format_cgw_items
 
 test_data_dir = os.path.join(os.path.dirname(__file__), "test_data")
 
@@ -16,3 +16,10 @@ def test_yaml_parser():
 
 def test_sort_items(fixture_unsorted_files_json, fixture_sorted_files_json):
     assert fixture_sorted_files_json == sort_items(fixture_unsorted_files_json)
+
+
+def test_yml_data_formats(yml_json_data):
+    yaml_file = os.path.join(test_data_dir, "test_yml_format.yaml")
+    cgw_items = yaml_parser(yaml_file)
+    formatted_cgw_data = format_cgw_items(cgw_items)
+    assert formatted_cgw_data == yml_json_data