+++ docs/CHANGELOG.rst
@@ -1,7 +1,39 @@
 ChangeLog
 =========
 
-0.1 (2022-15-09)
+0.3.0 (2022-16-11)
+------------------
+
+* Pubtools CGW can support nested structure of data in YAML file
+* Changed the YAML field `state` to `action` to denote more appropriate operation
+* File `order` in YAML will be auto incremented by 10 for nested data structure
+* Updated documentation
+
+
+0.2.3 (2022-24-11)
+------------------
+
+* Get push item size when proceses
+
+
+0.2.2 (2022-10-11)
+------------------
+
+* Additional integration fixes
+
+0.2.1 (2022-07-11)
+------------------
+
+* Fixed entry point loading for push-staged
+
+0.2.0 (2022-07-11)
+------------------
+
+* Integration of pubtools push-staged
+* Added documentation
+
+
+0.1.0 (2022-15-09)
 ------------------
 
 * Initial release.